+++ packages/platform-sdk-nano/src/dto/transactions/transfer.ts
@@ -1,0 +1,9 @@
+import { Contracts, Exceptions } from "@arkecosystem/platform-sdk";
+
+import { TransactionData } from "../transaction";
+
+export class TransferData extends TransactionData implements Contracts.TransferData {
+	public memo(): string | undefined {
+		return undefined;
+	}
+}