+++ packages/platform-sdk-nano/src/dto/wallet.ts
@@ -7,39 +7,41 @@ export class WalletData extends DTO.AbstractWalletData implements Contracts.Wall
 	}
 
 	public address(): string {
-		return "TODO";
+		return this.data.id;
 	}
 
 	public publicKey(): string | undefined {
-		throw new Exceptions.NotImplemented(this.constructor.name, "publicKey");
+		return undefined;
 	}
 
 	public balance(): Contracts.WalletBalance {
+		const available: BigNumber = BigNumber.make(this.data.balance).divide(1e30).times(1e8);
+
 		return {
-			total: BigNumber.ZERO,
-			available: BigNumber.ZERO,
-			fees: BigNumber.ZERO,
+			available,
+			fees: available,
+			locked: BigNumber.make(this.data.pending).divide(1e30).times(1e8),
 		};
 	}
 
 	public nonce(): BigNumber {
-		throw new Exceptions.NotImplemented(this.constructor.name, "nonce");
+		return BigNumber.ZERO;
 	}
 
 	public secondPublicKey(): string | undefined {
-		throw new Exceptions.NotImplemented(this.constructor.name, "secondPublicKey");
+		return undefined;
 	}
 
 	public username(): string | undefined {
-		throw new Exceptions.NotImplemented(this.constructor.name, "username");
+		return undefined;
 	}
 
 	public rank(): number | undefined {
-		throw new Exceptions.NotImplemented(this.constructor.name, "rank");
+		return undefined;
 	}
 
 	public votes(): BigNumber | undefined {
-		throw new Exceptions.NotImplemented(this.constructor.name, "votes");
+		return undefined;
 	}
 
 	public entities(): Contracts.Entity[] {