+++ packages/platform-sdk-xrp/src/dto/wallet.ts
@@ -16,7 +16,6 @@ export class WalletData extends DTO.AbstractWalletData implements Contracts.Wall
 
 	public balance(): Contracts.WalletBalance {
 		return {
-			total: BigNumber.make(this.data.balance).times(1e8),
 			available: BigNumber.make(this.data.balance).times(1e8),
 			fees: BigNumber.make(this.data.balance).times(1e8),
 		};